cosmos-mcp/requirements.txt
azure-cosmos
mcp
fastapi
uvicorn